✨ 💡 Add flag to pass full path to Envoy.blade.php file
This feature will allow users to override the expectaction that Envoy.blade.php is in the root of the project by supplying the full path on the system to Envoy.blade.php. Alternatively this could be implemented where they give just a path to the folder contianing the file. The current implementation expects the path and the filename.